docs(key-bindings): update widget descriptions (#19)
Commit b8c6b9a corresponds to the widget name change introduced in
https://github.com/atuinsh/atuin/pull/1631. Also, the current
documentation diverges from the current implementation. I updated the
widgets for the <kbd>up</kbd> key in commit 6e1f3df. Also, the
descriptions for the newly added widgets for `keymap_mode`
(https://github.com/atuinsh/atuin/pull/1570) are added in commit
705214a.

Since the version dependence of documentation can confuse people as in
https://github.com/atuinsh/atuin/issues/1625, I decided to add a note on
the version dependence and also a required version for each feature. I
personally think it is useful to maintain in the documentation the
minimal version requirement for each feature, but this might be the
perspective of a developer who needs to consider the compatibility of
products with arbitrary versions of other products. If you have a
different preference, please tell me that.

There is an extra commit 33a1d6d. Sorry, this is unrelated to the widget
name but a tiny clarification of a description I added in my previous PR
#10. If I should submit a change in a separate PR, please tell me that.
It's a small change, but I'll create a separate one if you'd prefer it.