Add documentation for Xonsh import (#29)
Documentation for Xonsh importers, as discussed in atuinsh/atuin#1678
and atuinsh/atuin#1711.

I also noticed that the TOC wasn't generating properly on the page, so I
changed the h1's to h2's and that seems to have fixed it. This does
slightly change the appearance of the page, so I'm happy to revert that
if you'd prefer to keep it the way it was.