docs(config): clarify scope of directory filter_mode
Currently the table leaves the impression that the directory mode is
limited to the current host, since from top to bottom is global, host,
and session. Session is obviously limited to the host, thus leaves the
next filter a bit ambiguous. Is it also limited to the host?

This change clarifies the scope of the filter.