docs(installation): add missing version requirement for ble.sh
We had mentioned the version requirement for ble.sh >= 0.4 in README
[1], but it was later removed without updating the atuinsh/docs
repository [2].  Currently, the information about the version
requirement for ble.sh is not found anywhere in the documentation.
This is a situation similar to Ref. [3].  This issue is found in a
discussion on forum.atuin.sh [4].  Here, we shall update the
documentation to include the change in Ref. [1].

References:

[1] https://github.com/atuinsh/atuin/pull/1465
[2] https://github.com/atuinsh/atuin/pull/2203
[3] https://github.com/atuinsh/docs/pull/64
[4] https://forum.atuin.sh/t/1047